pyproject.toml
[tool.poetry]
name = "gitops-utils"
version = "1.1.0"
description = ""
authors = ["Jon Bogaty <jon@jonbogaty.com>"]
readme = "README.md"
packages = [{include = "gitops_utils"}]

[tool.poetry.dependencies]
python = "^3.12"
case-insensitive-dictionary = "^0.2.1"
python-hcl2 = "^4.3.2"
ruamel-yaml = "^0.18.6"
orjson = "^3.9.15"
numpy = "^1.26.4"
deepmerge = "^1.1.1"
sortedcontainers = "^2.4.0"
more-itertools = "^10.2.0"
filelock = "^3.13.3"
inflection = "^0.5.1"
validators = "^0.24.0"
requests = "^2.31.0"
gitpython = "^3.1.42"
flatdict = "^4.0.1"
rich = "^13.7.1"
annotated-types = "^0.6.0"
pathvalidate = "^3.2.0"
werkzeug = "^3.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.1.1"
autoimport = "^1.4.0"
isort = "^5.13.2"
coverage = {extras = ["toml"], version = "^7.4.4"}
flake8 = "^7.0.0"
flake8-bugbear = "^24.2.6"
mypy = "^1.9.0"


[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
skip_gitignore = "true"
line_length = 88
profile = "black"

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 100


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
